requirements.txt
ipython==8.12.0
langchain
openai
pinecone-client
python-dotenv==1.0.0
uuid
pathlib
uvicorn
fastapi
jinja2
slowapi
httpx
git+https://github.com/casics/nostril.git
tiktoken
langsmith
bs4
llama_index
nltk